fix(connections): general tab rename
