+++ packages/connect-form/src/components/advanced-options-tabs/general-tab.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ConnectionStringUrl from 'mongodb-connection-string-url';
 
-import SchemaInput from './general/schema-input';
+import SchemaInput from './general-tab/schema-input';
 import { UpdateConnectionFormField } from '../../hooks/use-connect-form';
 import { ConnectionFormError } from '../../utils/connect-form-errors';
 import FormFieldContainer from '../form-field-container';
-import HostInput from './general/host-input';
+import HostInput from './general-tab/host-input';
 
 function GeneralTab({
   errors,
